\name{tglm-package}
\alias{tglm-package}
\alias{tglm}
\docType{package}
\title{
Binary Regressions under Independent Student-t Priors
}
\description{
Use Gibbs sampler with Polya-Gamma data augmentation to fit logistic and probit regression under independent Student-t priors (including Cauchy priors and normal priors as special cases). 
}
\details{
\tabular{ll}{
Package: \tab tglm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-07-24\cr
License: \tab GPL(>= 2)\cr
}
Main function: \code{\link{tglm.fit}}.
}
\author{
Yingbo Li

Maintainer: Yingbo Li <ybli@clemson.edu>
}
\references{
James H. Albert and Siddhartha Chib. (1993) "Bayesian Analysis of Binary and Polychotomous Response Data." Journal of the American statistical Association, 88(422), 669-679.

Nicholas G. Polson, James G. Scott, and Jesse Windle. (2013) "Bayesian Inference for Logistic Models Using Polya-Gamma Latent Variables." Journal of the American statistical Association, 108(504), 1339-1349. 

Joyee Ghosh, Yingbo Li, and Robin Mitra. "On the Use of Cauchy Prior Distributions for Bayesian Logistic Regression." Working paper. 
}
\keyword{ package }
\keyword{ logistic regression }
\keyword{ probit regression }
\keyword{ Cauchy prior }
\keyword{ MCMC }
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
